requirements.txt
requests==2.28.1
pytest==7.2.0
allure-pytest==2.12.0
